Re #1828: Fixed misc compile warnings (on iOS).
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@5035 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjmedia/src/pjmedia/sdp_neg.c
@@ -84,7 +84,7 @@ static pj_status_t custom_fmt_match( pj_pool_t *pool,
  */
 PJ_DEF(const char*) pjmedia_sdp_neg_state_str(pjmedia_sdp_neg_state state)
 {
-    if (state >=0 && state < (pjmedia_sdp_neg_state)PJ_ARRAY_SIZE(state_str))
+    if ((int)state >=0 && state < (pjmedia_sdp_neg_state)PJ_ARRAY_SIZE(state_str))
 	return state_str[state];
 
     return "<?UNKNOWN?>";
